Continua Desarrollo Negocio

+++ Dominio/Compra.cs
@@ -8,7 +8,7 @@ namespace Dominio
 {
     public class Compra
     {
-        public int IdCompra { get; set; }
+        public Int64 IdCompra { get; set; }
         public Proveedor Proveedor { get; set; }
         public Usuario Usuario { get; set; }
         public DateTime Fecha { get; set; }
